9_Desafio2/ServicesRepo/calculadora-metros-cuadrados/src/main/java/com/mercadolibre/calculadorametroscuadrados/controller/CalculateRestController.java
package com.mercadolibre.calculadorametroscuadrados.controller;

import com.mercadolibre.calculadorametroscuadrados.dto.HouseDTO;
import com.mercadolibre.calculadorametroscuadrados.dto.HouseResponseDTO;
import com.mercadolibre.calculadorametroscuadrados.exceptions.NotFoundLocation;
import com.mercadolibre.calculadorametroscuadrados.service.CalculateService;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;

@RestController
public class CalculateRestController {

  private final CalculateService calculateService;

  public CalculateRestController(CalculateService calculateService) {
    this.calculateService = calculateService;
  }

  @PostMapping("/calculate")
  public HouseResponseDTO calculate(@RequestBody @Valid HouseDTO house) throws NotFoundLocation {
    return calculateService.calculate(house);
  }

}
